requirements.txt
asgiref==3.8.1
Django==4.2.15
django-autoslug==1.9.9
django-cities-light==3.10.1
django-environ==0.11.2
progressbar2==4.4.2
python-utils==3.8.2
pytz==2024.1
sqlparse==0.5.1
typing_extensions==4.12.2
Unidecode==1.3.8